rename solar orbit module
